Move export methods to controller
