+++ src/controller/controller.py
@@ -6,6 +6,9 @@ from src.store.project_information_store import update_project_store
 from PyQt5.QtWidgets import *
 from src.store.project_information_store import project_information_store
 import json
+from src.domain.exports import export_g6_to_png, export_g6_to_tikz, export_g6_to_pdf
+from src.view.loading.loading_window import LoadingWindow
+from PyQt5 import QtCore
 
 
 class Controller:
@@ -15,6 +18,8 @@ class Controller:
         self.wizard_controller = WizardController()
         self.filter_controller = FilterController()
         self.project_controller = ProjectController()
+
+        self.loading_window = LoadingWindow()
 
         self.current_open_window = ""
 
@@ -35,6 +40,11 @@ class Controller:
     def connect_project_events(self):
         self.project_controller.project_window.new_action.triggered.connect(self.show_wizard_window)
         self.project_controller.project_window.open_action.triggered.connect(self.show_open_project_window)
+
+        self.project_controller.project_window.export_png_action.triggered.connect(self.export_to_png)
+        self.project_controller.project_window.export_g6_action.triggered.connect(self.export_to_g6)
+        self.project_controller.project_window.export_tikz_action.triggered.connect(self.export_to_tikz)
+        self.project_controller.project_window.export_pdf_action.triggered.connect(self.export_to_pdf)
 
     def connect_events(self):
         self.connect_welcome_events()
@@ -102,3 +112,61 @@ class Controller:
 
     def close_project_window(self):
         self.wizard_controller.close_window()
+
+    def get_name_from_save_dialog(self, format_file):
+        file_name = QFileDialog.getSaveFileName(parent=self.project_controller.project_window,
+                                                caption=self.project_controller.project_window.tr(f"Export graphs to {format_file} file"),
+                                                filter=self.project_controller.project_window.tr(f"Text files (*.{format_file})"),
+                                                directory=f"{project_information_store.project_name}.{format_file}"
+                                                )[0]
+        if file_name:
+            if not QtCore.QFileInfo(file_name).suffix():
+                file_name += f".{format_file}"
+        return file_name
+
+    def export_to_png(self):
+        file_dir = str(QFileDialog.getExistingDirectory(parent=self.project_controller.project_window, caption="Select Directory"))
+        if file_dir:
+            self.show_loading_window(len(project_information_store.filtered_graphs))
+            for step, graph in enumerate(project_information_store.filtered_graphs):
+                export_g6_to_png(graph, file_dir, step)
+                self.update_loading_window(step)
+            self.loading_window.close()
+
+    def export_to_tikz(self):
+        file_dir = str(QFileDialog.getExistingDirectory(parent=self.project_controller.project_window, caption="Select Directory"))
+        if file_dir:
+            self.show_loading_window(len(project_information_store.filtered_graphs))
+            for step, graph in enumerate(project_information_store.filtered_graphs):
+                export_g6_to_tikz(graph, file_dir, step)
+                self.update_loading_window(step)
+            self.loading_window.close()
+
+    def export_to_pdf(self):
+        file_dir = str(QFileDialog.getExistingDirectory(parent=self.project_controller.project_window, caption="Select Directory"))
+        if file_dir:
+            self.show_loading_window(len(project_information_store.filtered_graphs))
+            for step, graph in enumerate(project_information_store.filtered_graphs):
+                export_g6_to_pdf(graph, file_dir, step)
+                self.update_loading_window(step)
+            self.loading_window.close()
+
+    def export_to_g6(self):
+        file_name = self.get_name_from_save_dialog('g6')
+        if file_name:
+            self.show_loading_window(len(project_information_store.filtered_graphs))
+            file = open(file_name, 'w')
+            for step, graph in enumerate(project_information_store.filtered_graphs):
+                file.write(graph)
+                self.update_loading_window(step)
+            file.close()
+            self.loading_window.close()
+
+    def show_loading_window(self, set_total):
+        self.loading_window.progressBar.setMaximum(set_total)
+        self.loading_window.progressBar.setValue(0)
+        self.loading_window.show()
+
+    def update_loading_window(self, step):
+        self.loading_window.increase_step(step)
+        QApplication.processEvents()
